requirements.txt
Django==1.9.1
#psycopg2==2.6.1
requests==2.6.0
simplejson==3.6.5
wsgiref==0.1.2
djangorestframework
markdown
django-filter
django-cors-headers
djangorestframework-jwt
